Make the vertex-remap LUT an int one
take's out= is the mesh's own int index column, so gathering from a float
table casts across dtype kinds. It is exact for indices this side of
2**53, and these meshes carry a few million, but numpy 2.5 deprecates it
and a later one will refuse -- and this runs twice per ground removal, so
every pgs-remove-ground-plane invocation printed the warning. The sibling
LUT in keep_triangles_by_mask was already int.

+++ pgs_recon/utils/geometry.py
@@ -343,8 +343,10 @@ def keep_vertices_by_mask(mesh: Mesh, mask):
     keep_triangles_by_mask(mesh, tri_mask)
 
     # LUT for (vid + None,) -> new_vid
-    # Unlike other mesh properties, vid's should never be None
-    lut = np.full((len_v,), -1.)
+    # Unlike other mesh properties, vid's should never be None.
+    # Int, because take's out= below is the mesh's own index column: gathering
+    # floats into it casts across dtype kinds, which numpy 2.5 deprecates
+    lut = np.full((len_v,), -1, dtype=int)
     new_idx = np.arange(mask.nonzero()[0].shape[0], dtype=int)
     lut[mask] = new_idx
     v_map = mesh.faces[..., 0].astype(int)
